feat: merge test files
